requirements.txt
streamlit
pandas
plotly
requests
google-generativeai